feat(API GraphQL public queries): add the 'isInTheCart' field
Added support for the 'isInTheCart' field in these GraphQL queries:
- product(id: Int!): Product
- featuredProducts: [Product]!
- products(page: Int!): GetProductByPageReturnValue!

These queries are available to the general public, they don't have to
be admins to access them. However, if they request the 'isInTheCart'
field, they have to provide the access token that's associated with
their account. Otherwise they will receive an error saying
'User must be authenticated to request the "isInTheCart" field'.

This change is needed for displaying products on the frontend. If some
product is in the cart, a different behaviour may be applied in this
case.

The API documentation has been modified to reflect this new change.