feat(API cartRedis model): create an additional overload for 'cleanCart'
This commits adds a new overload for the 'cleanCart' function inside
the 'cartRedis' model. Previously, this function was responsible for
cleaning the cart of one user. However, it's possible that an upcoming
feature will require to delete multiple Redis keys at once. Instead of
calling this function for each key we want to delete, a new overload
was implemented that allows to pass multiple user IDs as an array of
numbers. This way, only one request containing multiple keys will be
sent to the Redis server instead of many requests for each id.

This change will potentially reduce the load on the Redis server.