fix(API GraphQL deleteProduct): invalidate Redis cart cache for affected users
When a product is deleted using the deleteProduct GraphQL mutation,
some users may have this product in their cart, and this cart may be
cached in Redis. When the product is deleted, it will be automatically
deleted from PostgreSQL, along with any cart records that contain this
product. However, the automatic deletion is only done for PostgreSQL,
not for Redis. This can lead to inconsistencies between the cart data
in PostgreSQL and the cart data in Redis.

This commit fixes this by completely removing the cart cache for all
users who had the deleted product id in their cached cart. This action
is performed in the background, the admin user won't need to wait for
the execution of this operation.

Q: "Why are we completely deleting the cart cache, but not directly
modifying it in Redis to exclude the deleted product?"
A: Imagine we have 100,000 users who have the deleted product in their
cart cached in Redis. We would need to get the cart (load it into RAM),
find the deleted product, remove it, and then set the cart in Redis.
And all this for EVERY user. This would consume a lot of resources on
the API server and the Redis server. This approach doesn't scale well.
It's much better to simply invalidate the cart cache for such users,
and if they use the cart again, it will be automatically cached in
Redis with consistent data.