fix(API GraphQL updateProduct): invalidate Redis cart cache for affected users
Similar to the previous commit, this commit fixes the cart cache
invalidation problem while using the updateProduct GraphQL mutation.

When a product is updated using the updateProduct mutation, some users
may have this product in their cart, and this cart may be cached in
Redis. Redis stores some basic information about cart products: title,
price, initialImageUrl etc. When a product is updated, the updated
information goes to PostgreSQL, potentially making the data in Redis
stale. One way to fix it would be to get those users who were affected
and for each user get their cart, modify it, and put it back into Redis.
However, this approach would require making a lot of requests to the
Redis server, which could overwhelm it. A much simpler approach would
be to get a list of affected users and invalidate their cart cache in
Redis. This is done by making 1 request to PostgreSQL and 1 request to
Redis.

It's also important to note that the execution time of the
updateProduct mutation won't degrade, because this cart invalidation
process occurs in the background. When the product is updated in
PostgreSQL, the resolver immediately returns, and the cart invalidation
process is initiated, which will happen in the background.