fix(React <EditProduct />): prevent duplicate updateProduct API request
When the product is updated using the <EditProduct /> component, the
API request made to the updateProduct GraphQL mutation was made twice,
which doubled the amount of time the user had to wait before seeing
the success message. This commit fixes this bug. Now the update product
request is only made once.