fix(React <AddProduct />): refetch GET_FEATURED_PRODUCTS after adding a product
With this commit, when a product is added by an admin using the
<AddProduct /> component, the GET_FEATURED_PRODUTS GraphQL query will
be refetched from the API server.

Strictly speaking, this change is not necessary for simple users, but
for admins this is quite a useful feature, because without it, the
products that are displayed on the main page may miss the added product.