fix(React Apollo Client): include __typename when overriding cache
Previously, whenever the updateProduct mutation was executed, cache
of some queries was explicitly overridden to reflect the new changes.
Here's a list of these queries:
- adminProduct (get details about a product using admin credentials)
- product (information about a specific product)
- featuredProducts (shown on the main page of the React app)

When the cache of these queries was updated, it didn't include the
__typename field, resulting in a warning in the console. The __typename
field is used by Apollo Client for internal caching operations, so it's
better to preserve it.