fix(React): invalidate GET_PRODUCTS_BY_PAGE GraphQL cache upon product deletion
When a product is deleted by an admin, the local cache of the
GET_PRODUCTS_BY_PAGE GraphQL queries isn't refetched, making it stale
(the deleted product was still there). The GET_PRODUCTS_BY_PAGE query
is used to fetch products by specifying a page.

This commit fixes this problem by removing the stale cache.

It's important to note that the affected queries won't be refetched
until the user requires it. For example, the query will be sent to the
server only if a user goes to "/products?page=N".