fix(React): invalidate GET_PRODUCTS_BY_PAGE GraphQL cache when adding a product
When a product is added by an admin, the GET_PRODUCTS_BY_PAGE cache
needs to be deleted, because this cache may not contain the newly added
product, making this cache potentially invalid.

This commit deletes any cache associated with the getProductsByPage
GraphQL query. It's also important to note that Apollo Client will not
try to refetch any queries unless the user goes to a page that actually
requires this getProductsByPage query. This way, the cache will be
deleted without impacting the performance of the API server.