feat(React): update cart icons on products based on cart status
This commit introduces changes to the cart icons displayed in two
scenarios:

- On the main page (via the featuredProducts GraphQL query).
- On various pages, such as /products?page=2, utilizing the
"products($page: Int!)" query.

In both cases, the underlying GraphQL queries now incorporate the
"isInTheCart" field, which signifies whether an item exists in the cart
or not.

Consequently, the "Add to cart" icon, visible on each product card,
dynamically adapts its appearance in response to the item's cart
status. If the item is in the cart, the icon reflects this; otherwise,
it presents a different icon.