feat(React): add the ability to add ONE item to the cart
Every product has a <AddToCartButton /> component attached to it.
Previously, this component didn't do anything, but this commit enhances
this component by adding the ability to add one product to the cart. If
a user clicks on this <AddToCartButton /> component the first time, it
adds this product to the cart. If the user clicks this button again,
they will be able to see the contents of their cart (right now showing
cart contents is not implemeneted). If the user tries to click the
button while the request is pending, nothing will happen because the
button click won't be handled.

Notes:
- It's possible to use the <AddToCartButton /> component on the main
page (using the "Explore Our Products" block) and using the "products
by page" functionality. This is because this <AddToCartButton />
component is used within the <ProductItem /> component.
- When a product is added to the cart, the "My Cart" button is
highlighted (a short animation is played to attract the user's
attention).
- When a product is added to the cart, any local product cache is
explicitly modified. Product refetching using the API doesn't happen at
all. BUT it does refetch the number of items in the cart using the
GET /api/user/cart/count endpoint.