fix(React <Logout />): abort Logout request on component unmount
Previously, when running the React app in the development mode, the
<Logout /> component showed an error when trying to log out. This is
because the logout API request was sent twice, and the API server has
implemented a rate limiting mechanism on the logout route, so sending
the request twice caused the nginx server to reply with the 503 status
code.

The request was sent twice because the React app uses the StrictMode in
development, so all useEffects are run twice (even with an empty deps
array).

This commit fixes this bug by providing a cleanup function to the
useEffect that sends the logout request to the server. This cleanup
function aborts the previous logout request.

While this fix isn't necessary for the production environment since
StrictMode doesn't apply to production, it still enhances the stability
of the React app.