refactor(React <Modal />): disable/enable body scroll inside the Modal component
Previously, if you wanted to utilize the <Modal /> component and you
wanted to implement a body scroll lock, you had to write the
implementation yourself for each component that utilizes the Modal
component.

This commit refactors the Modal component to implement the body scroll
lock/unlock logic out of the box. This greatly simplifies the logic of
many components that depend on the <Modal /> component. Now if you
mount the Modal component, the body scroll is locked automatically,
and if the component unmounts, it's unlocked also automatically.

Components that depend on the <Modal /> component were also refactored
to remove unnecessary logic related to the body scroll lock/unlock
logic because it's no longer needed.