refactor(React <ProductInfo />): create a separate component for quantity selector
The ProductInfo component is used to display product information on a
single page. This component renders another component called
ProductDescription, which, in turn, renders the ProductCartActions
component. The ProductCartActions component contains a utility field
that allows to specify some number using the '+' (increase quantity)
and '-' sign (decrease quantity). The logic for this field was
implemented entirely in the ProductCartActions component. However,
there is now a need to reuse this quantity selector field in another
place in the application. It's needed to specify how many instances of
a single product the user wants to add to the cart. This is why this
commit implements a separate component for this quantity selector. This
new component is called <QuantitySelector />

Possible props for this component:
- initialValue - value that will be shown when the component mounts
- onQuantityChange - a function that will be called every time the
current number changes within this field. This function will be
provided with the current quantity as an argument. This function will
not be called when the component initially mounts.