refactor(React <Modal />): separate the backdrop from the modal
Previously, the div block responsible for the modal window was inside
the backdrop block. However, this approach turned out to be suboptimal,
because it adds tight coupling to these components. For example, when
adding a click event listener to document.body, when the user tried to
click on the modal window, it didn't propagate to document.body,
because the backdrop used event.stopPropagation(). It wasn't possible
to allow the propagation, because otherwise the modal would close
itself after the user clicked on any element within the modal window.

This is why it was decided to separate the backdrop and the modal
window. Now they are sibling elements.