feat(GET /api/user/cart): include totalPrice field in the response
Previously, when users accessed their cart via the GET /api/user/cart
endpoint, they only received a list of products. However, the React.js
application now requires the total price of the products as well. This
commit addresses this requirement by enhancing the endpoint to
calculate the total price of the cart products on the server side. The
total price is now included in the response as a 'totalPrice' field.
This enhancement eliminates the need for clients to calculate the price
individually. The updated response from the GET /api/user/cart endpoint
is as follows:

{
    products: {
        productId: number;
        title: string;
        price: number;
        initialImageUrl: string;
        quantity: number;
    },
    totalPrice: number
}

The API documentation was also modified to reflect this change.