fix(API productModel.getProduct()): convert 'price' to a number
When a user tried to fetch the contents of their cart, sometimes the
'price' field displayed as a string. This was because of the way the
'getProduct' method of the product model handled the 'price' field.
This method fetched the price field from PostgreSQL without converting
it to a number first. This commit fixes this bug. Now the getProduct
method ensures the 'price' field will always be a number. And this, in
turn, ensures that when a user fetches the cart, the 'price' field will
always be displayed as a number.