feat(React <QuantitySelector />): allow to specify a range of values
This commit enhances the QuantitySelector component by making it adjust
the current value that is displays. This component now accepts 2 new
prop values:
- minValue - minimum value that can be displayed inside the
QuantitySelector component
- maxValue - maximum value that can be displayed
The component will never show value less or greater than the defined
boundaries. This was done by implementing a new custom hook called
"useNumberRange".

This change is necessary for implementing the cart functionality. The
user will be able to select how many products they want to add to their
cart. It's necessary to ensure that the values are within a defined
range of values.

Note:
The <ProductCartActions /> component was slightly modified to reflect
this new feature. This component now displays values ranging from 1 to
32767 (because you can add up to 32767 instances of the same product to
the cart).