refactor(React): rename RTK Query API hook for clarity
RTK Query allows you to create hooks for different API endpoints.
One of these endpoints allows you to insert new products into the cart
or update the quantity of existing products. Previously, the hook
that's associated with this endpoint was called "pushToCart". However,
this name doesn't accurately represent the function's behavior. This
commit renames it to a more suitable name "upsertCartProduct". It
clearly conveys the meaning of this operation.

The AddToCartButton component utilized the old name of the hook, so
this component was modified to use the new hook name.