feat(React): simplify the useDebounce hook to reduce the number of invocations
Previously, the useDebounce hook invoked the provided function
immediately if this is the first time executing the function (this
behavior is called the leading edge), or if the difference between the
time of the last invocation and the current time exceeds the specified
threshold (the second argument of useDebounce). Otherwise the execution
is scheduled. This way, we got this behavior:
- This first execution of the debounced function immediately invokes
the provided function
- The next debounced invocation is scheduled
- All the subsequent debounced invocations will cancel the previous
scheduled invocation
- If you try to call the debounced function and the difference between
the time of the last invocation and the current time exceeds the
specified threshold, then the function is called immediately, without
any scheduling.

To avoid confusion and to simplify this process, it was decided to
rewrite the useDebounce logic:
- The first debounced call is executed immediately, without scheduling,
just like before
- The next call is scheduled
- The next call cancels the previous scheduled call, and is scheduled
to be executed after the specified delay
- If you call the debounced function after a long time, it won't be
executed immediately, but rather scheduled, just like any other call

This simplification brings two advantages:
- The entire flow is easier to understand
- The number of calls is reduced. This way, you may get less calls to
the Node.js API if your debounced function calls some API endpoint.