feat(React useDebounce): add the 'leadingEdge' option
This commit adds the 'leadingEdge' option when calling the useDebounce
hook. This option allows to specify whether the debounced function
should be executed immediately without any scheduling when it's first
called. The default is false.
- If 'leadingEdge' is set to true, the initial call will be executed
immediately, and then all the subsequent calls will be scheduled. Every
call cancels the previous scheduled call, but the first call can't be
canceled because it's executed immediately.
- If 'leadingEdge' is set to false (default), all calls will be
scheduled, and every call cancels the previous scheduled call.

This change is required because previously the useDebounce hook was
implemented to use 'leadingEdge' by default, and there was no way to
change it. Currently, a new feature is being implemented that allows
users to add products to the cart. This feature requires the
useDebounce hook with the 'leadingEdge' option set to false.

Additional note:
The useSignupValidation hook was slighly modified, because it uses the
useDebounce hook. This hook was modified to set the 'leadingEdge'
option to true.