fix(React <QuantitySelector />): prevent calling props.onQuantityChange on mount
Even though the QuantitySelector component was initially built to
prevent the props.onQuantityChange function from being called on
component mount, it was still called because of React.StrictMode, which
causes components to mount, unmount, and then mount again. Although
this behavior only occurs in the development, it's still frustrating
and needs to be addressed.

This commit fixes this bug by modifying the useNumberRange hook to
expose the "isValueChanged" property. The useNumberRange hook returns
the current number and different methods to modify this value by
incrementing it, decrementing, or setting a fixed value for it.
In short, this hook is like an advanced reusable counter. The newly
added "isValueChanged" property specifies whether the value was changed
in any way (whether the value was incremented/decremented or if the
value was explicitly set by the "setValue" function).