feat(React useDebounce): add the ability to cancel a scheduled function
The useDebounce hook is used to debounce a function call. This commit
adds a new feature to this custom hook. Now this hook returns a new
function called "cancelDebouncedExecution", which cancels the scheduled
function execution.

This feature is needed to implement the UI for working with the cart.
Use cases for "cancelDebouncedExecution":
- A user tries to increment the product quantity in the cart, and then
immediately goes back to the previous value, even before the API call
is executed to modify the cart product quantity. In this case it's
necessary to cancel the debounced API call, which prevents any
unnecessary requests.
- A user tries to add too many products to the cart (for example there
are 20 product in stock, but the user is trying to add 50 products to
the cart). In this case the React application will try to roll back the
product quantity to the previous valid value. When the rollback is
done, the API call is canceled because there's no need to send it.