feat(React <Image />): add the 'invertColor' prop
The <Image /> component is used to display a small decorative icon.
It's usually used for displaying icons in front of navigation links.
Previously, this component inverted the color of the icon and there was
no way to change it. This commit adds a new prop to the Image component
called "invertColor", which allows to invert the color of the icon. By
default it's set to true.

This change is needed to implement the UI for working with carts. If a
user's cart is empty, a special icon will be displayed. This icon
doesn't need color inversion, so this commit adds the new prop to the
<Image /> component just for this purpose.