feat(React): add the ability to show the cart
This commit allows users to see the contents of their cart, as well as
manipulate the items inside of it. These manupulations include:
- Adding/Subtracting cart items using the QuantitySelector component
- Completely deleting a product from the cart

Users can access their cart either by adding an item to the cart, and
then clicking on the "add to cart" button again, or by clicking on the
"My Cart" component in the upper right corner of the page.

If a user tries to add too many products to the cart (for example, if
the user is trying to add more products than are available in stock),
then the product quantity will be rolled back to the last valid value,
and an appropriate error notification will be shown.