fix(React product info page): remove QuantitySelector component
Reminder: <QuantitySelector /> is used to display an input along with
the 'plus' and 'minus' button to manipulate the current value displayed
on the screen. It's typically used to manipulate the number of items
in the cart.

In one of previous commits, the <QuantitySelector /> component was
changed to accept different methods to manipulate the current value
instead of modifying it inside the QuantitySelector component itself.
Because of this change, the product info page was broken. This commit
removes the broken functionality. The reason behind deleting it is that
right now there's no need for this functionality. In the future, the
QuantitySelector component will be replaced.