refactor(React): move all GraphQL queries/mutations to one location
Previously, all GraphQL queries and mutations were scattered across
the whole React project directory. The location depended on where these
queries/mutations were used. For example, if the "getFeaturedProducts"
query was used inside the <ExploreProductsBlock /> component, then the
file that defines the GraphQL query was defined inside the same folder
as the ExploreProductsBlock component.

The problem with this approach is that multiple components can use the
same query or mutation. Another reason is that it's just easier to work
with GraphQL queries and mutations if they are located in one place.

This is why now all GraphQL queries and mutations are located inside
/client/src/graphql/queries and /client/src/graphql/mutations folders
respectively.

Affected components were modified to update import paths to these
queries and mutations.