feat(React <ProductInfo />): allow users to add a product to the cart
Allows users to add a product to their cart from the product info page
(e.g. /products/1096). Users can now add the product to their cart by
clicking on the "Add to cart" button. If the product wasn't it the
cart, the button shows one icon. If a product is added to the cart or
was already added, the button shows another icon.

When the user adds a product to the cart via the product info page, the
local product cache is updated such that this change is visible in
other places without triggering an API request. For example, if the
user has the "featuredProducts" GraphQL query response in their cache,
adding a product to the cart will modify this cache so that the
"isInTheCart" property is set to true. This won't trigger any requests
to the REST API.