feat(docker-compose.yaml): add pgadmin4 as a service
This commit adds a new service to 'docker-compose.yaml', which is a
docker compose configuration file for development purposes. This new
service is called 'pgadmin', which allows to interact with the Postgres
RDS instance using GUI.

For this new service to work properly, the 'dev.env' file is necessary.
This file should be located as '/pg-admin/dev.env'. It should contain
these environment variables:
- PGADMIN_DEFAULT_EMAIL
- PGADMIN_DEFAULT_PASSWORD
These variables define credentials which are used to log in to pgAdmin.

The docker compose file was configured to map the container port 80 to
the host port 5000. This way, you can now access pgAdmin by typing
"http://localhost:5000".