feat(docker-compose.yaml): add postgres as a service
This will be useful as a sandbox for experiments. For example, to try
applying migrations without worrying about breaking anything.

The PostgreSQL container exposes a port 5432 to the host machine and it
uses a configuration file called /postgres/dev.env. This file defines
only one parameter POSTGRES_PASSWORD="..."

You can log in to this DB using the user "postgres" and typing in the
password that you defined in POSTGRES_PASSWORD.