fix(React <ProductInfo />): show an appropriate message if a product wasn't found
Previously, when a user tried to visit a page with information about a
product that doesn't exist (e.g. http://localhost/products/532461266),
the React app showed an infinite loading screen. This bug is fixed
with this commit. Now when a user tries to get info about a non-existent
product, they will get an appropriate message, namely "A product with
the specified id does not exist"