fix(React): refetch the cart for admins if a product is deleted
If an admin account adds a product to the cart, and then immediately
deletes this product from the whole application with the 'Delete'
button, the cart should be automatically updated to reflect this change.
However, it didn't happen, the cart still showed a non-existent product
after the product deletion. This commit fixes this bug. Now if an admin
deletes a product, their cart will be refetched from the API.