refactor(React): use a <LoadingScreen /> component to display a loading state
Many components in the React app use this JSX snippet:
<div className="flex-wrapper">
    <Loading />
</div>
This shows a loading spinner in the center of the page.

To avoid duplication across the React application, this JSX snippet
was put inside a new component called "LoadingScreen". Now every
component can use it instead of copy-pasting the same JSX code over and
over again.