refactor(API): add a surrogate key to 'product_categories'
Previously, the 'product_categories' table only had one primary key
called 'category', whose type is VARCHAR(30). However, it's generally
recommended to use numeric columns as primary keys because they're more
efficient for indexing, sorting, and querying. Another reason for this
is that using a VARCHAR column as a primary key would mean that any
foreign keys associated with this primary key would use a lot of
storage, so it's much more efficient to use a numeric surrogate key as
it would consume less storage.

This commit creates a new migration file that changes the DB schema to
create a surrogate numeric auto-incrementing 'id' key for the
'product_categories' table. The seeding files and the app functionality
were modified to reflect this change and prevent the app from breaking.