fix(API getProductCategories controller): handle async errors properly with asyncHandler
Previously, if an unexpected error occurred in the getProductCategories
controller method while a user was making a request, an error would be
thrown internally, but the user's request would just hang. This commit
fixes this by wrapping the controller method with 'asyncHandler'. Now,
if the promise is rejected, Express will catch this error and the user
will get an error message "Something went wrong".