feat(API GraphQL 'products' query)!: add support for product categories
Previously, there was a GraphQL query called "products" that allowed
users to fetch products by specifying a page. However, this query
didn't allow to specify a product category, which is why the API
returned all kinds of products by specifying a page.

This commit modifies this GraphQL query to allow to also specify the
'category' value as an argument.

Now this query is called this way (as an example):

query Products($category: String!, $page: Int!) {
  products(category: $category, page: $page) {
    productList {
      id
      title
      price
      category
      initialImageUrl
      additionalImageUrl
      shortDescription
      isAvailable
      isRunningOut
      isInTheCart
    }
    totalPages
  }
}

Specified variables:
{
  "category": "Ноутбуки",
  "page": 1
}

This query can return the same errors as before, namely:
- The 'page' parameter is less than or equal to zero.
- The user is trying to access the 'isInTheCart' field while being
unauthenticated

In addition to these errors, this commit also added handling for
another error that is associated with product categories:
- The user specified a non-existent product category

The API documentation was also updated to reflect these changes.

BREAKING CHANGE:
The React.js application must be modified to display products by
categories.