feat(React): display products by categories
When a user clicks on the "Products" menu item, a page opens up in
front of them, showing all available product categories. When the user
selects a category, various products belonging to the selected category
will be shown. This commit adds support for displaying products
belonging to different categories. Pagination is also implemented to
display only a couple of products rather than all of them.