feat(React.js main menu): display <MyCartButton /> and <SignInOrSignUp /> buttons to anonymous users
Preciously, anonymous users would see the "SignIn" or "SignUp" buttons
in the top right corner of the main menu. Right now there's a need to
implement an ability for anonymous users to interact with the cart.
However, this is only possible if the "My Cart" button is shown on the
main menu for anonymous users. This commit displays this button, as
well as a new button with the title "Sign In / Sign Up" instead of two
buttons "Sign In" and "Sign Up". This was done to make sure there are
only 2 buttons shown instead of 3.

Currently, clicking the "My Cart" button triggers an error as the React
app attempts to request cart items from the API, which is only allowed
for authenticated users. Future commits will address this by
implementing cart functionality using Redux and/or localStorage. This
approach ensures local storage of cart data, enabling seamless
interaction even for anonymous users. When users log in, their locally
stored cart data will be sent to the server for persistence. Similarly,
if an anonymous user attempts to create an order, the locally stored
cart data, along with their information, will generate an account
storing the previously local cart products. Note that these features
are not yet implemented, this is only to understand the context of this
commit.