feat(API): add an endpoint to check if it's safe to add a product to the local cart
Context:
When an anonymous user tries to modify the number of products in their
local cart (which is stored in localStorage), the React application
needs a way to check whether there are enough products in stock and
whether the user is allowed to order the specified number of products.
For authenticated users it's done with the API request that adds a
product to the cart. However, anonymous users can't access this
endpoint, so there's a need for a new endpoint that will allow
anonymous users to find out whether it's safe to add the provided
quantity of products to the cart. This commit implements this endpoint
and adds a detailed overview of it to the API documentation.

New endpoint:
GET /api/user/cart/is-safe-to-add-product?productId=155&quantityToAdd=6

Purpose:
Checks whether there are enough products and whether one user can add
the specified quantity of products to their local cart.
Basically, it works like this:
- The user sends a request to this endpoint, specifying 'productId' and
'quantityToAdd' (the quantity of one product that they want to add to
the cart) as query parameters.
- The server pulls out the 'quantity_in_stock' and 'max_order_quantity'
fields from the requested product and then compares it to the specified
'quantityToAdd' parameter. If everything is alright, it returns this
response:
200 OK; { "safeToAdd": true, "reason": null }
If the user is trying to add more products than is in stock:
200 OK; { "safeToAdd": false, "reason": "InsufficentProductStock" }
If the user is trying to add more products than allowed:
200 OK; { "safeToAdd": false, "reason": "ExceededMaxOrderQuantity",
"maxOrderQuantity": 2 }

Possible errors:
- The provided product is not a number (422 Unprocessable Entity)
- The provided quantityToAdd parameter is not a positive number or is
not a number at all (422 Unprocessable Entity)

Please refer to the API documentation for more detailed information.