refactor(React): decompose <CartProduct /> into smaller components
The CartProduct component became too hard to manage because of its size.
This commit splits this component into much smaller components that do
only one function. For example, deleting a cart product is delegated to
the CartProductDeleteButton component, which is part of the
AdditionalOptions component, which, in turn, is part of the CartProduct
component. Modifying the exact quantity of an item in the cart is done
using the CartProductQuantity component. This way we get a much clearer
separation of concerns, which makes the cart much more manageable.