fix(React): hide any error notifications when changing the cart product quantity
Previously, when the user changed the quantity of some product in the
cart, a user could see an error notification at the top left side of
the screen. However, when the user tried to change the quantity again,
the previous error wouldn't hide. This could confuse the user. This
commit addresses this issue by making sure any errors are hidden
whenever the user changes cart product quantity. However, errors won't
be hidden if a quantity rollback has happened. For example, a user has
20 instances of some product, and then tries to increase it to 100000,
the API will indicate that it's impossible to add that much products
because they aren't in stock, so the React app will display an error
notification, and then roll back to the last known cart product
quantity. This rollback, however, won't hide the recently showed error,
because it's relevant to the user.