refactor(API): move 'isProductInTheCart' to the cart model
The 'isProductInTheCart' function is responsible for checking whether a
specified user has the specified product in their cart. Previously,
this function was in the product model. However, this commit changes
this by moving this function to the cart model instead. This was done
to improve code organization, as well as increase readability and
discoverability. This refactoring contributes to a cleaner, more
maintainable codebase.