feat(API & React): enforce a maximum quantity of cart products
This commit introduces a new limitation to the ability of users to add
products to their cart (PUT /api/user/cart endpoint). Now, if a user is
trying to add more products than "process.env.MAX_PRODUCTS_IN_CART",
the server will return an error. It's important to mention that these
products must be unique.
For example, if process.env.MAX_PRODUCTS_IN_CART = 3, and the user is
trying to add the 4th product to their cart, the API server will return
an error. But if the user has 3 unique products in the cart and tries
to change the quantity of one of them, the server will fulfill the
request without any questions.

The React application has been modified to show this new error if it
occurs. The API documentation has also been revised to reflect this
change.